src/main/java/com/ms/easy/FindFirstRepeatingCharacter.java
package com.ms.easy;

import java.util.LinkedHashMap;
import java.util.Map;
import java.util.function.Function;
import java.util.stream.Collectors;

public class FindFirstRepeatingCharacter {

    public static Character findFirstRepeatingCharacter(String input){
        return input.chars()
                .mapToObj(i -> Character.toLowerCase((char) i))
                .collect(Collectors.groupingBy(Function.identity(), LinkedHashMap::new, Collectors.counting()))
                .entrySet().stream()
                .filter(entry -> entry.getValue() > 1)
                .map(Map.Entry::getKey)
                .findFirst().orElseThrow(() -> new RuntimeException("No Repeating Character"));
    }
}
